Nett: Remove WithRoot as it is a broken premature optimization
The method was weird anyway as it somewhat did the same as clone
but not exactly. The idea was that WithRoot creates a new
TOML object only with the real data but not other complex object graph
parts like comments.

But in reality this apporoach only broke functionality and nothing good.

If perf really is an issue at some point probably techniques like
a copy on write collection for the comments etc. is a better approach.

Related: #84